+++ internal/manifests/collector/parser/helpers_test.go
@@ -1,0 +1,64 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package parser
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestComponentType(t *testing.T) {
+	for _, tt := range []struct {
+		desc     string
+		name     string
+		expected string
+	}{
+		{"regular case", "myreceiver", "myreceiver"},
+		{"named instance", "myreceiver/custom", "myreceiver"},
+	} {
+		t.Run(tt.desc, func(t *testing.T) {
+			// test and verify
+			assert.Equal(t, tt.expected, ComponentType(tt.name))
+		})
+	}
+}
+
+func TestReceiverParsePortFromEndpoint(t *testing.T) {
+	for _, tt := range []struct {
+		desc          string
+		endpoint      string
+		expected      int
+		errorExpected bool
+	}{
+		{"regular case", "http://localhost:1234", 1234, false},
+		{"absolute with path", "http://localhost:1234/server-status?auto", 1234, false},
+		{"no protocol", "0.0.0.0:1234", 1234, false},
+		{"just port", ":1234", 1234, false},
+		{"no port at all", "http://localhost", 0, true},
+	} {
+		t.Run(tt.desc, func(t *testing.T) {
+			// test
+			val, err := portFromEndpoint(tt.endpoint)
+			if tt.errorExpected {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			assert.EqualValues(t, tt.expected, val, "wrong port from endpoint %s: %d", tt.endpoint, val)
+		})
+	}
+}
