+++ internal/manifests/collector/podmonitor.go
@@ -80,18 +80,7 @@ func PodMonitor(params manifests.Params) (*monitoringv1.PodMonitor, error) {
 }
 
 func metricsEndpointsFromConfig(logger logr.Logger, otelcol v1beta1.OpenTelemetryCollector) []monitoringv1.PodMetricsEndpoint {
-	// TODO: https://github.com/open-telemetry/opentelemetry-operator/issues/2603
-	cfgStr, err := otelcol.Spec.Config.Yaml()
-	if err != nil {
-		logger.V(2).Error(err, "Error while marshaling to YAML")
-		return []monitoringv1.PodMetricsEndpoint{}
-	}
-	config, err := adapters.ConfigFromString(cfgStr)
-	if err != nil {
-		logger.V(2).Error(err, "Error while parsing the configuration")
-		return []monitoringv1.PodMetricsEndpoint{}
-	}
-	exporterPorts, err := adapters.ConfigToComponentPorts(logger, adapters.ComponentTypeExporter, config)
+	exporterPorts, err := adapters.ConfigToComponentPorts(logger, adapters.ComponentTypeExporter, otelcol.Spec.Config.Exporters.Object)
 	if err != nil {
 		logger.Error(err, "couldn't build endpoints to podMonitors from configuration")
 		return []monitoringv1.PodMetricsEndpoint{}
